Backend: Modified /auth/v1/refresh endpoint
I have setup the frontend to make a http request to the
/auth/v1/refresh endpoint as soon as it is loaded in the browser. While
trying to integrate the refresh endpoint in the frontend, i noticed
that my refresh endpoint logic didn't work correctly if the browser was
refreshed several times in quick succession. The refresh logic didn't
work correctly because in the first request it would mark the status of
the received 'valid' status valued jwt refresh token as 'invalidated' and then further successive requests would send the same previously 'invalidated' status
valued jwt refresh token and then all of the user's refresh tokens
would be set to status value of 'compromised' and a 401 no body
response message would be sent back to the client instead of getting a
jwt access token and a new jwt refresh token cookie. In order to solve
this problem i decided to add a extension period of 5 minutes to each
jwt refresh token with status value of 'invalidated'. During this
extension period of 5 minutes even jwt refresh tokens with status of
'invalidated' are allowed to produce a jwt access token and set a new
jwt refresh token, so basically they are treated like having status of
'valid' during the extension period. I created a new table called invalidated_jwt_refresh_token_extension_periods that has a one to one
relationship with my jwt_refresh_tokens table. Whenever a 'valid'
status valued jwt refresh token is marked as 'invalidated', a row that
points to that 'invalidated' token is added in the new table and a
timestamp column called extension_period_ends_at is also stored in that
row that has value that is 5 minutes in the future from when the jwt
refresh token is marked with status value of 'invalidated'. So
basically the new table contains the info about the extension period of
the 'invalidated' status valued jwt refresh tokens.