Frontend: Display todos fetched from GET /api/v1/todos?<query-string> endpoint in the home page's TodoArea component
The content of the fetched todos which is in quill text editor's
supported delta format is displayed in a read only no toolbar quill
text editor while the title is displayed in a <input type="text">
element. Three span elements are also shown for each todo for the
purposes of editing,deleting and displaying info respectively.