Backend: Modify http request message cookie logic in /auth/v1/logout endpoint
Http request message cookie logic used to stop the logout process
if the Cookie request message header contained more than 1 cookie.
Now it doesn't do that and just checks if a cookie key matches
the name expected for containing the jwt refresh token and if it
does then it extracts the value and continues with the logout process
whereas if no cookie key is found that matches the expected name then
the logout process is aborted. The change was made as more cookies
could possibly be added in future and therefore multiple cookies within
the Cookie http request message header shouldn't be rejected.