Backend: Modify 'invalidated' status valued jwt refresh token handling logic in /auth/v1/refresh endpoint
Earlier the logic assumed that every jwt refresh token with a status
value of 'invalidated' had an associated row in the invalidated_jwt_refresh_token_extension_periods table. But this
assumption is wrong as the /auth/v1/logout endpoint changes the
status value from 'valid' to 'invalidated' for 'valid' status
valued jwt refresh tokens without inserting an associated row in the
invalidated_jwt_refresh_token_extension_periods table as logging out
implies that the recieved jwt refresh token won't be sent again
from the legitimate frontend client so it doesn't require an
extension period. So while checking if a invalidated status
valued jwt refresh token's extension period is expired there is a
possibility that no row will be found in the
invalidated_jwt_refresh_token_extension_periods table and thus,
the logic was modified to account for this.